+++ app/models.py
@@ -2,6 +2,7 @@
 from pydantic import BaseModel, HttpUrl
 
 
+# #UFB-0006
 class URLMessage(BaseModel):
     url: HttpUrl
     is_group_chat: bool = False